expand_rownames: expand parameter descriptions in rownames

Move rownames that describe parameters (e.g. xi[2009]) to columns.

expand_rownames(x, time_name, geo_name, group_names)

Arguments

x
A table with rownames in the format param[group1__groupK,t] or param[t].

time_name
A name for any resulting time variable.

geo_name
A name for any resulting geographic variable.

group_names
Names for any resulting group variables.

Value

x with additional columns (see details).

Details

It should rarely be necessary to call expand_rownames directly. But elements extracted from dgirtfit-class objects may have rownames of the format param[group1__groupK,t] for parameters indexed by group and time period, or param[t] for parameters indexed by time period. expand_rownames moves this information to columns whose names are given by the col_names argument. The rownames in their original format will appear in another column called rn.

See also

dgirtfit-class