Class dgirt_fit: a class for fitted models

dgirt returns a fitted model object of class dgirt_fit, which inherits from dgo_fit.

Details

dgo 0.2.8 deprecated the dgirtfit class and replaced it with the dgirt_fit class.

Slots

dgirt_in

dgirtin-class data used to fit the model.

See also

dgmrp_fit dgo_fit

Examples

data(toy_dgirtfit)
# summarize the fitted results
summary(toy_dgirtfit, pars = 'xi')

#> dgirt samples from 4 chains of 400 iterations, 200 warmup, thinned every 1 
#>   Drawn Mon May 29 16:21:17 2017 
#>   Package version 0.2.10 
#>   Model version 2017_01_04 
#>   43 parameters; 12 theta_bars (year state race3)
#>   2 periods 2009 to 2010 
#> 
#> n_eff

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   26.92  168.29  245.19  279.58  330.50  800.00 

#> 
#> Rhat

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>  0.9985  1.0065  1.0115  1.0141  1.0161  1.1075 

#> 
#> Elapsed time

#>    chain warmup sample total
#> 1:     1     2S     3S    5S
#> 2:     2     3S     4S    7S
#> 3:     3     2S     2S    4S
#> 4:     4     3S     2S    5S

# get posterior means with a convenience function
get_posterior_mean(toy_dgirtfit, pars = 'theta_bar')

#>         param state race3 year         mean
#>  1: theta_bar    SC black 2009  2.581303889
#>  2: theta_bar    SC black 2010  1.332893651
#>  3: theta_bar    SC other 2009 -1.474896258
#>  4: theta_bar    SC other 2010  0.002222642
#>  5: theta_bar    SC white 2009 -1.862723209
#>  6: theta_bar    SC white 2010 -1.563808392
#>  7: theta_bar    VA black 2009  2.305825333
#>  8: theta_bar    VA black 2010  1.637632813
#>  9: theta_bar    VA other 2009 -1.095120475
#> 10: theta_bar    VA other 2010  0.300496502
#> 11: theta_bar    VA white 2009 -1.078641503
#> 12: theta_bar    VA white 2010 -0.887081709

# generally apply functions to posterior samples after warmup; n.b.
# `as.array` is iterations x chains x parameters so `MARGIN = 3` applies
# `FUN` over iterations and chains
apply(as.array(toy_dgirtfit, pars = 'xi'), 3, mean)

#>      xi[1]      xi[2] 
#> 1.64610835 0.02308406 

# access the posterior samples
head(as.data.frame(toy_dgirtfit, pars = 'theta_bar'))

#>        param state race3 year iteration     value
#> 1: theta_bar    SC black 2009         1 1.0253959
#> 2: theta_bar    SC black 2009         2 0.8703493
#> 3: theta_bar    SC black 2009         3 2.5546916
#> 4: theta_bar    SC black 2009         4 2.8851330
#> 5: theta_bar    SC black 2009         5 1.6683215
#> 6: theta_bar    SC black 2009         6 6.4071224