print method for dgo_fit-class objects

print method for dgo_fit-class objects get_elapsed_time: extract chain run times from dgo_fit-class objects summary method for dgo_fit-class objects summarize method for dgo_fit-class objects as.data.frame method for dgo_fit-class objects rhats: extract split R-hats from dgo_fit-class objects

# S4 method for dgo_fit
show(object)

# S4 method for dgo_fit
print(x, ...)

print.dgo_fit(x, ...)

# S4 method for dgo_fit
get_elapsed_time(object, ...)

# S4 method for dgo_fit
summary(object, ..., verbose = FALSE)

# S4 method for dgo_fit
get_posterior_mean(object, pars = "theta_bar", ...)

summarize(x, ...)

# S4 method for dgo_fit
summarize(x, pars = "theta_bar", funs = c("mean", "sd",
  "median", "q_025", "q_975"))

# S3 method for dgo_fit
as.data.frame(x, ..., pars = "theta_bar",
  keep.rownames = FALSE)

rhats(x, ...)

# S4 method for dgo_fit
rhats(x, pars = "theta_bar")

Arguments

object
A dgo_fit-class object

x
A dgo_fit-class object

...
Further arguments to stanfit-class methods.

verbose
Whether to show the full output from the rstan method.

pars
Parameter name(s)

funs
Quoted names of summary functions. `q_025` is accepted as shorthand for `function(x) quantile(x, .025)`, and similarly `q_975`.

keep.rownames
Whether to retain original parameter names with numeric indexes, as output from RStan.

Value

A table giving split R-hats for model parameters

Examples

data(toy_dgirtfit)
summarize(toy_dgirtfit)

#>         param state race3 year         mean        sd       median      q_025
#>  1: theta_bar    SC black 2009  2.581303889 2.0773702  2.025744919  0.4663683
#>  2: theta_bar    SC black 2010  1.332893651 0.9342120  1.087193518  0.3108696
#>  3: theta_bar    SC other 2009 -1.474896258 1.5187934 -1.073520925 -5.1898485
#>  4: theta_bar    SC other 2010  0.002222642 0.8380635 -0.001426992 -1.7762351
#>  5: theta_bar    SC white 2009 -1.862723209 1.5588242 -1.463595482 -5.9613244
#>  6: theta_bar    SC white 2010 -1.563808392 1.1325962 -1.296122850 -4.4077860
#>  7: theta_bar    VA black 2009  2.305825333 1.9368960  1.756194045  0.4316253
#>  8: theta_bar    VA black 2010  1.637632813 1.1228341  1.317204830  0.4178312
#>  9: theta_bar    VA other 2009 -1.095120475 1.2038704 -0.784426840 -4.2372739
#> 10: theta_bar    VA other 2010  0.300496502 0.6713402  0.227488241 -0.7974740
#> 11: theta_bar    VA white 2009 -1.078641503 0.9660087 -0.836360911 -3.8903031
#> 12: theta_bar    VA white 2010 -0.887081709 0.6906443 -0.705485848 -2.6783883
#>           q_975
#>  1:  8.70794851
#>  2:  3.74446230
#>  3:  0.01442099
#>  4:  1.58650903
#>  5: -0.26197625
#>  6: -0.25046472
#>  7:  7.49825382
#>  8:  4.62354118
#>  9:  0.24311882
#> 10:  1.79960198
#> 11: -0.12456895
#> 12: -0.08468892

data(toy_dgirtfit)
# access posterior samples
as.data.frame(toy_dgirtfit, pars = 'theta_bar')

#>           param state race3 year iteration      value
#>    1: theta_bar    SC black 2009         1  1.0253959
#>    2: theta_bar    SC black 2009         2  0.8703493
#>    3: theta_bar    SC black 2009         3  2.5546916
#>    4: theta_bar    SC black 2009         4  2.8851330
#>    5: theta_bar    SC black 2009         5  1.6683215
#>   ---                                                
#> 9596: theta_bar    VA white 2010       796 -0.9399185
#> 9597: theta_bar    VA white 2010       797 -0.8616181
#> 9598: theta_bar    VA white 2010       798 -0.5838702
#> 9599: theta_bar    VA white 2010       799 -0.6810680
#> 9600: theta_bar    VA white 2010       800 -0.2927825

data(toy_dgirtfit)
rhats(toy_dgirtfit)

#> Error in summary(x, par = pars, verbose = TRUE)$summary: $ operator is invalid for atomic vectors