print method for dgo_fit-class objects

print method for dgo_fit-class objects get_elapsed_time: extract chain run times from dgo_fit-class objects summary method for dgo_fit-class objects summarize method for dgo_fit-class objects as.data.frame method for dgo_fit-class objects rhats: extract split R-hats from dgo_fit-class objects

# S4 method for dgo_fit
show(object)

# S4 method for dgo_fit
print(x, ...)

print.dgo_fit(x, ...)

# S4 method for dgo_fit
get_elapsed_time(object, ...)

# S4 method for dgo_fit
summary(object, ..., verbose = FALSE)

# S4 method for dgo_fit
get_posterior_mean(object, pars = "theta_bar", ...)

summarize(x, ...)

# S4 method for dgo_fit
summarize(x, pars = "theta_bar", funs = c("mean", "sd",
  "median", "q_025", "q_975"))

# S3 method for dgo_fit
as.data.frame(x, ..., pars = "theta_bar",
  keep.rownames = FALSE)

rhats(x, ...)

# S4 method for dgo_fit
rhats(x, pars = "theta_bar")

Arguments

object
A dgo_fit-class object

x
A dgo_fit-class object

...
Further arguments to stanfit-class methods.

verbose
Whether to show the full output from the rstan method.

pars
Parameter name(s)

funs
Quoted names of summary functions. `q_025` is accepted as shorthand for `function(x) quantile(x, .025)`, and similarly `q_975`.

keep.rownames
Whether to retain original parameter names with numeric indexes, as output from RStan.

Value

A table giving split R-hats for model parameters

Examples

data(toy_dgirtfit)
summarize(toy_dgirtfit)

#>         param state race3 year        mean        sd      median      q_025
#>  1: theta_bar    SC black 2009  2.35662560 1.6415237  2.02449197  0.5230645
#>  2: theta_bar    SC black 2010  1.22256665 0.8371110  1.06239289  0.2941792
#>  3: theta_bar    SC other 2009 -1.36379358 1.1174329 -1.10842179 -3.9993518
#>  4: theta_bar    SC other 2010  0.04227331 0.8781789  0.05571854 -1.6000637
#>  5: theta_bar    SC white 2009 -1.63196374 1.1129132 -1.37631685 -4.5840865
#>  6: theta_bar    SC white 2010 -1.47489402 1.0881085 -1.27747059 -4.2971941
#>  7: theta_bar    VA black 2009  1.90274077 1.2776764  1.67871864  0.4932506
#>  8: theta_bar    VA black 2010  1.52339960 1.0105135  1.31034193  0.3788519
#>  9: theta_bar    VA other 2009 -0.95920551 1.0113154 -0.74479466 -3.3714601
#> 10: theta_bar    VA other 2010  0.28518022 0.6061495  0.22600882 -0.8581018
#> 11: theta_bar    VA white 2009 -0.92697350 0.6981114 -0.77099368 -2.7673600
#> 12: theta_bar    VA white 2010 -0.83419892 0.6838827 -0.69712692 -2.5838185
#>          q_975
#>  1:  6.6377822
#>  2:  3.4078614
#>  3: -0.0339191
#>  4:  1.7330404
#>  5: -0.3398887
#>  6: -0.2790527
#>  7:  5.1766964
#>  8:  4.1273832
#>  9:  0.1818680
#> 10:  1.7790140
#> 11: -0.1639113
#> 12: -0.1079362

data(toy_dgirtfit)
# access posterior samples
as.data.frame(toy_dgirtfit, pars = 'theta_bar')

#>           param state race3 year iteration      value
#>    1: theta_bar    SC black 2009         1  0.7183680
#>    2: theta_bar    SC black 2009         2  0.7609363
#>    3: theta_bar    SC black 2009         3  0.7942637
#>    4: theta_bar    SC black 2009         4  1.4685053
#>    5: theta_bar    SC black 2009         5  1.1297978
#>   ---                                                
#> 9596: theta_bar    VA white 2010       796 -1.7576671
#> 9597: theta_bar    VA white 2010       797 -0.9733774
#> 9598: theta_bar    VA white 2010       798 -0.9442844
#> 9599: theta_bar    VA white 2010       799 -1.7814752
#> 9600: theta_bar    VA white 2010       800 -2.3038391

data(toy_dgirtfit)
rhats(toy_dgirtfit)

#>         param state race3 year      Rhat
#>  1: theta_bar    SC black 2009 1.0231219
#>  2: theta_bar    VA black 2009 1.0156964
#>  3: theta_bar    SC other 2009 1.0180730
#>  4: theta_bar    VA other 2009 1.0102031
#>  5: theta_bar    SC white 2009 1.0221029
#>  6: theta_bar    VA white 2009 1.0170904
#>  7: theta_bar    SC black 2010 1.0099752
#>  8: theta_bar    VA black 2010 1.0097523
#>  9: theta_bar    SC other 2010 0.9957962
#> 10: theta_bar    VA other 2010 1.0027474
#> 11: theta_bar    SC white 2010 1.0125680
#> 12: theta_bar    VA white 2010 1.0104212