print method for dgo_fit-class objects

print method for dgo_fit-class objects get_elapsed_time: extract chain run times from dgo_fit-class objects summary method for dgo_fit-class objects summarize method for dgo_fit-class objects as.data.frame method for dgo_fit-class objects rhats: extract split R-hats from dgo_fit-class objects

# S4 method for dgo_fit
show(object)

# S4 method for dgo_fit
print(x, ...)

print.dgo_fit(x, ...)

# S4 method for dgo_fit
get_elapsed_time(object, ...)

# S4 method for dgo_fit
summary(object, ..., verbose = FALSE)

# S4 method for dgo_fit
get_posterior_mean(object, pars = "theta_bar", ...)

summarize(x, ...)

# S4 method for dgo_fit
summarize(x, pars = "theta_bar", funs = c("mean", "sd",
  "median", "q_025", "q_975"))

# S3 method for dgo_fit
as.data.frame(x, ..., pars = "theta_bar",
  keep.rownames = FALSE)

rhats(x, ...)

# S4 method for dgo_fit
rhats(x, pars = "theta_bar")

Arguments

object
A dgo_fit-class object

x
A dgo_fit-class object

...
Further arguments to stanfit-class methods.

verbose
Whether to show the full output from the rstan method.

pars
Parameter name(s)

funs
Quoted names of summary functions. `q_025` is accepted as shorthand for `function(x) quantile(x, .025)`, and similarly `q_975`.

keep.rownames
Whether to retain original parameter names with numeric indexes, as output from RStan.

Value

A table giving split R-hats for model parameters

Examples

data(toy_dgirtfit)
summarize(toy_dgirtfit)

#>         param state race3 year        mean        sd      median      q_025
#>  1: theta_bar    SC black 2009  2.24559756 1.7897209  1.84180053  0.4522434
#>  2: theta_bar    SC black 2010  1.15755259 0.9116460  0.88343031  0.2488049
#>  3: theta_bar    SC other 2009 -1.29813621 1.2173550 -0.99560220 -4.3745255
#>  4: theta_bar    SC other 2010  0.03000299 0.6902895  0.06824939 -1.4082368
#>  5: theta_bar    SC white 2009 -1.59792561 1.2283236 -1.31250037 -4.7712845
#>  6: theta_bar    SC white 2010 -1.29950834 1.0557300 -1.01959622 -4.0458618
#>  7: theta_bar    VA black 2009  2.02881527 1.4012611  1.67166246  0.4524494
#>  8: theta_bar    VA black 2010  1.40617435 1.0288018  1.11423319  0.3391811
#>  9: theta_bar    VA other 2009 -0.93038106 1.0475514 -0.69822446 -3.4956185
#> 10: theta_bar    VA other 2010  0.27922913 0.5149350  0.21708687 -0.6543708
#> 11: theta_bar    VA white 2009 -0.91333932 0.7394027 -0.72245960 -2.7489035
#> 12: theta_bar    VA white 2010 -0.71577875 0.6234497 -0.55060586 -2.3174531
#>           q_975
#>  1:  6.80052090
#>  2:  3.52798518
#>  3:  0.02768110
#>  4:  1.39749747
#>  5: -0.28580519
#>  6: -0.19539782
#>  7:  5.46631775
#>  8:  4.35076456
#>  9:  0.35390364
#> 10:  1.43108820
#> 11: -0.11261952
#> 12: -0.06558131

data(toy_dgirtfit)
# access posterior samples
as.data.frame(toy_dgirtfit, pars = 'theta_bar')

#>           param state race3 year iteration      value
#>    1: theta_bar    SC black 2009         1  2.3585309
#>    2: theta_bar    SC black 2009         2  3.7881672
#>    3: theta_bar    SC black 2009         3  2.5182470
#>    4: theta_bar    SC black 2009         4  4.9504883
#>    5: theta_bar    SC black 2009         5  0.5702558
#>   ---                                                
#> 9596: theta_bar    VA white 2010       796 -0.5245107
#> 9597: theta_bar    VA white 2010       797 -1.4022078
#> 9598: theta_bar    VA white 2010       798 -1.0872082
#> 9599: theta_bar    VA white 2010       799 -0.9778072
#> 9600: theta_bar    VA white 2010       800 -0.9032673

data(toy_dgirtfit)
rhats(toy_dgirtfit)

#>         param state race3 year      Rhat
#>  1: theta_bar    SC black 2009 0.9988403
#>  2: theta_bar    VA black 2009 1.0058328
#>  3: theta_bar    SC other 2009 1.0046187
#>  4: theta_bar    VA other 2009 1.0049422
#>  5: theta_bar    SC white 2009 1.0019199
#>  6: theta_bar    VA white 2009 1.0058461
#>  7: theta_bar    SC black 2010 1.0042491
#>  8: theta_bar    VA black 2010 1.0061799
#>  9: theta_bar    SC other 2010 1.0011256
#> 10: theta_bar    VA other 2010 1.0001827
#> 11: theta_bar    SC white 2010 1.0097541
#> 12: theta_bar    VA white 2010 1.0048440