print method for dgo_fit-class objects

print method for dgo_fit-class objects get_elapsed_time: extract chain run times from dgo_fit-class objects summary method for dgo_fit-class objects summarize method for dgo_fit-class objects as.data.frame method for dgo_fit-class objects rhats: extract split R-hats from dgo_fit-class objects

# S4 method for dgo_fit
show(object)

# S4 method for dgo_fit
print(x, ...)

print.dgo_fit(x, ...)

# S4 method for dgo_fit
get_elapsed_time(object, ...)

# S4 method for dgo_fit
summary(object, ..., verbose = FALSE)

# S4 method for dgo_fit
get_posterior_mean(object, pars = "theta_bar", ...)

summarize(x, ...)

# S4 method for dgo_fit
summarize(x, pars = "theta_bar", funs = c("mean", "sd",
  "median", "q_025", "q_975"))

# S3 method for dgo_fit
as.data.frame(x, ..., pars = "theta_bar",
  keep.rownames = FALSE)

rhats(x, ...)

# S4 method for dgo_fit
rhats(x, pars = "theta_bar")

Arguments

object
A dgo_fit-class object

x
A dgo_fit-class object

...
Further arguments to stanfit-class methods.

verbose
Whether to show the full output from the rstan method.

pars
Parameter name(s)

funs
Quoted names of summary functions. `q_025` is accepted as shorthand for `function(x) quantile(x, .025)`, and similarly `q_975`.

keep.rownames
Whether to retain original parameter names with numeric indexes, as output from RStan.

Value

A table giving split R-hats for model parameters

Examples

data(toy_dgirtfit)
summarize(toy_dgirtfit)

#>         param state race3 year        mean        sd      median      q_025
#>  1: theta_bar    SC black 2009  2.27918561 1.7765565  1.86898580  0.4389700
#>  2: theta_bar    SC black 2010  1.22043041 0.8070593  1.01803870  0.2795176
#>  3: theta_bar    SC other 2009 -1.39020024 1.3277101 -1.06220157 -4.6182915
#>  4: theta_bar    SC other 2010  0.04736177 0.7065802  0.06484053 -1.4736401
#>  5: theta_bar    SC white 2009 -1.67254198 1.2911376 -1.31200008 -4.3990168
#>  6: theta_bar    SC white 2010 -1.36955157 0.9607030 -1.12572336 -3.7612912
#>  7: theta_bar    VA black 2009  2.01678885 1.3905011  1.70467508  0.4775803
#>  8: theta_bar    VA black 2010  1.44993059 0.8939907  1.18348901  0.4264043
#>  9: theta_bar    VA other 2009 -0.98106291 1.0104605 -0.71838736 -3.7562032
#> 10: theta_bar    VA other 2010  0.31428591 0.5423229  0.24438855 -0.7620008
#> 11: theta_bar    VA white 2009 -0.94262607 0.7377999 -0.75788299 -2.6032035
#> 12: theta_bar    VA white 2010 -0.77981128 0.6057161 -0.61792883 -2.4385546
#>            q_975
#>  1:  5.859103945
#>  2:  3.428210162
#>  3:  0.006463068
#>  4:  1.529480693
#>  5: -0.275482452
#>  6: -0.241320122
#>  7:  5.201661089
#>  8:  3.901259857
#>  9:  0.218800853
#> 10:  1.587097574
#> 11: -0.108416797
#> 12: -0.104858051

data(toy_dgirtfit)
# access posterior samples
as.data.frame(toy_dgirtfit, pars = 'theta_bar')

#>           param state race3 year iteration      value
#>    1: theta_bar    SC black 2009         1  1.7567618
#>    2: theta_bar    SC black 2009         2  1.8128853
#>    3: theta_bar    SC black 2009         3  2.2679425
#>    4: theta_bar    SC black 2009         4  2.4373390
#>    5: theta_bar    SC black 2009         5  2.4654188
#>   ---                                                
#> 9596: theta_bar    VA white 2010       796 -1.0171975
#> 9597: theta_bar    VA white 2010       797 -1.0151980
#> 9598: theta_bar    VA white 2010       798 -0.8902516
#> 9599: theta_bar    VA white 2010       799 -0.8992006
#> 9600: theta_bar    VA white 2010       800 -0.8992006

data(toy_dgirtfit)
rhats(toy_dgirtfit)

#>         param state race3 year     Rhat
#>  1: theta_bar    SC black 2009 1.010926
#>  2: theta_bar    VA black 2009 1.020005
#>  3: theta_bar    SC other 2009 1.007219
#>  4: theta_bar    VA other 2009 1.004017
#>  5: theta_bar    SC white 2009 1.043383
#>  6: theta_bar    VA white 2009 1.032628
#>  7: theta_bar    SC black 2010 1.022345
#>  8: theta_bar    VA black 2010 1.023130
#>  9: theta_bar    SC other 2010 1.001243
#> 10: theta_bar    VA other 2010 1.001773
#> 11: theta_bar    SC white 2010 1.025877
#> 12: theta_bar    VA white 2010 1.026677