print method for dgo_fit-class objects

print method for dgo_fit-class objects get_elapsed_time: extract chain run times from dgo_fit-class objects summary method for dgo_fit-class objects summarize method for dgo_fit-class objects as.data.frame method for dgo_fit-class objects rhats: extract split R-hats from dgo_fit-class objects

# S4 method for dgo_fit
show(object)

# S4 method for dgo_fit
print(x, ...)

print.dgo_fit(x, ...)

# S4 method for dgo_fit
get_elapsed_time(object, ...)

# S4 method for dgo_fit
summary(object, ..., verbose = FALSE)

# S4 method for dgo_fit
get_posterior_mean(object, pars = "theta_bar", ...)

summarize(x, ...)

# S4 method for dgo_fit
summarize(x, pars = "theta_bar", funs = c("mean", "sd",
  "median", "q_025", "q_975"))

# S3 method for dgo_fit
as.data.frame(x, ..., pars = "theta_bar",
  keep.rownames = FALSE)

rhats(x, ...)

# S4 method for dgo_fit
rhats(x, pars = "theta_bar")

Arguments

object
A dgo_fit-class object

x
A dgo_fit-class object

...
Further arguments to stanfit-class methods.

verbose
Whether to show the full output from the rstan method.

pars
Parameter name(s)

funs
Quoted names of summary functions. `q_025` is accepted as shorthand for `function(x) quantile(x, .025)`, and similarly `q_975`.

keep.rownames
Whether to retain original parameter names with numeric indexes, as output from RStan.

Value

A table giving split R-hats for model parameters

Examples

data(toy_dgirtfit)
summarize(toy_dgirtfit)

#>         param state race3 year        mean        sd     median      q_025
#>  1: theta_bar    SC black 2009  2.24053065 1.9336147  1.7799611  0.4566614
#>  2: theta_bar    SC black 2010  1.27549292 1.0036580  0.9944347  0.2931024
#>  3: theta_bar    SC other 2009 -1.40019273 1.6534406 -1.0144666 -4.7212136
#>  4: theta_bar    SC other 2010  0.04084414 0.7356635  0.0503198 -1.4892224
#>  5: theta_bar    SC white 2009 -1.67983508 1.6607712 -1.2961321 -4.8426620
#>  6: theta_bar    SC white 2010 -1.42533238 1.1166382 -1.1603282 -4.4630400
#>  7: theta_bar    VA black 2009  2.01710000 1.7614584  1.6346355  0.4384231
#>  8: theta_bar    VA black 2010  1.48835032 1.0317741  1.1923152  0.3542094
#>  9: theta_bar    VA other 2009 -1.02559194 1.1759219 -0.6812453 -4.2629804
#> 10: theta_bar    VA other 2010  0.30256578 0.6388466  0.2265883 -0.9024376
#> 11: theta_bar    VA white 2009 -0.93711250 0.9046802 -0.7106767 -3.1100210
#> 12: theta_bar    VA white 2010 -0.82327290 0.6901749 -0.6491099 -2.6735110
#>           q_975
#>  1:  6.45595603
#>  2:  3.83169121
#>  3:  0.06443927
#>  4:  1.63704312
#>  5: -0.23570371
#>  6: -0.23587321
#>  7:  5.41852042
#>  8:  4.01645824
#>  9:  0.16390533
#> 10:  1.77105059
#> 11: -0.07141857
#> 12: -0.08582941

data(toy_dgirtfit)
# access posterior samples
as.data.frame(toy_dgirtfit, pars = 'theta_bar')

#>           param state race3 year iteration       value
#>    1: theta_bar    SC black 2009         1  1.81338340
#>    2: theta_bar    SC black 2009         2  2.18738932
#>    3: theta_bar    SC black 2009         3  4.87691903
#>    4: theta_bar    SC black 2009         4  1.63898360
#>    5: theta_bar    SC black 2009         5  3.59430508
#>   ---                                                 
#> 9596: theta_bar    VA white 2010       796 -0.32568849
#> 9597: theta_bar    VA white 2010       797 -0.12337371
#> 9598: theta_bar    VA white 2010       798 -0.08819301
#> 9599: theta_bar    VA white 2010       799 -0.27716272
#> 9600: theta_bar    VA white 2010       800 -0.03302348

data(toy_dgirtfit)
rhats(toy_dgirtfit)

#>         param state race3 year      Rhat
#>  1: theta_bar    SC black 2009 1.0342446
#>  2: theta_bar    VA black 2009 1.0316121
#>  3: theta_bar    SC other 2009 1.0167080
#>  4: theta_bar    VA other 2009 1.0179279
#>  5: theta_bar    SC white 2009 1.0335336
#>  6: theta_bar    VA white 2009 1.0370154
#>  7: theta_bar    SC black 2010 1.0253885
#>  8: theta_bar    VA black 2010 1.0292079
#>  9: theta_bar    SC other 2010 0.9971425
#> 10: theta_bar    VA other 2010 1.0011172
#> 11: theta_bar    SC white 2010 1.0287917
#> 12: theta_bar    VA white 2010 1.0217522