print method for dgo_fit-class objects

print method for dgo_fit-class objects get_elapsed_time: extract chain run times from dgo_fit-class objects summary method for dgo_fit-class objects summarize method for dgo_fit-class objects as.data.frame method for dgo_fit-class objects rhats: extract split R-hats from dgo_fit-class objects

# S4 method for dgo_fit
show(object)

# S4 method for dgo_fit
print(x, ...)

print.dgo_fit(x, ...)

# S4 method for dgo_fit
get_elapsed_time(object, ...)

# S4 method for dgo_fit
summary(object, ..., verbose = FALSE)

# S4 method for dgo_fit
get_posterior_mean(object, pars = "theta_bar", ...)

summarize(x, ...)

# S4 method for dgo_fit
summarize(x, pars = "theta_bar", funs = c("mean", "sd",
  "median", "q_025", "q_975"))

# S3 method for dgo_fit
as.data.frame(x, ..., pars = "theta_bar",
  keep.rownames = FALSE)

rhats(x, ...)

# S4 method for dgo_fit
rhats(x, pars = "theta_bar")

Arguments

object
A dgo_fit-class object

x
A dgo_fit-class object

...
Further arguments to stanfit-class methods.

verbose
Whether to show the full output from the rstan method.

pars
Parameter name(s)

funs
Quoted names of summary functions. `q_025` is accepted as shorthand for `function(x) quantile(x, .025)`, and similarly `q_975`.

keep.rownames
Whether to retain original parameter names with numeric indexes, as output from RStan.

Value

A table giving split R-hats for model parameters

Examples

data(toy_dgirtfit)
summarize(toy_dgirtfit)

#>         param state race3 year       mean        sd      median      q_025
#>  1: theta_bar    SC black 2009  2.3366278 1.7170886  1.86230582  0.5726695
#>  2: theta_bar    SC black 2010  1.2377753 0.9192781  0.98003214  0.2923020
#>  3: theta_bar    SC other 2009 -1.3572482 1.2744087 -1.02682180 -4.5653631
#>  4: theta_bar    SC other 2010  0.0488835 0.7121616  0.05180635 -1.4820937
#>  5: theta_bar    SC white 2009 -1.6299612 1.2284926 -1.33245202 -4.7810038
#>  6: theta_bar    SC white 2010 -1.3977524 1.1435021 -1.12326373 -4.4516855
#>  7: theta_bar    VA black 2009  2.0786160 1.4494523  1.70271518  0.5200655
#>  8: theta_bar    VA black 2010  1.4686529 1.0529409  1.18578363  0.3615049
#>  9: theta_bar    VA other 2009 -1.0419862 1.0771621 -0.76291860 -3.6955714
#> 10: theta_bar    VA other 2010  0.2789239 0.5709138  0.23354275 -0.7495978
#> 11: theta_bar    VA white 2009 -0.9452116 0.7729931 -0.75014128 -2.8037034
#> 12: theta_bar    VA white 2010 -0.8000586 0.7410834 -0.62790473 -2.6776297
#>            q_975
#>  1:  7.178353073
#>  2:  3.742430465
#>  3: -0.005582985
#>  4:  1.625935250
#>  5: -0.319337381
#>  6: -0.229948533
#>  7:  5.848106088
#>  8:  4.168408318
#>  9:  0.186711949
#> 10:  1.641921654
#> 11: -0.123163301
#> 12: -0.071849680

data(toy_dgirtfit)
# access posterior samples
as.data.frame(toy_dgirtfit, pars = 'theta_bar')

#>           param state race3 year iteration      value
#>    1: theta_bar    SC black 2009         1  3.1165518
#>    2: theta_bar    SC black 2009         2  4.5381575
#>    3: theta_bar    SC black 2009         3  1.3696296
#>    4: theta_bar    SC black 2009         4  2.9999388
#>    5: theta_bar    SC black 2009         5  1.7527522
#>   ---                                                
#> 9596: theta_bar    VA white 2010       796 -0.4810858
#> 9597: theta_bar    VA white 2010       797 -0.5445991
#> 9598: theta_bar    VA white 2010       798 -1.0540017
#> 9599: theta_bar    VA white 2010       799 -0.8035109
#> 9600: theta_bar    VA white 2010       800 -0.7178630

data(toy_dgirtfit)
rhats(toy_dgirtfit)

#>         param state race3 year      Rhat
#>  1: theta_bar    SC black 2009 1.0150177
#>  2: theta_bar    VA black 2009 1.0201624
#>  3: theta_bar    SC other 2009 1.0116201
#>  4: theta_bar    VA other 2009 1.0111787
#>  5: theta_bar    SC white 2009 1.0184768
#>  6: theta_bar    VA white 2009 1.0219520
#>  7: theta_bar    SC black 2010 1.0154753
#>  8: theta_bar    VA black 2010 1.0193238
#>  9: theta_bar    SC other 2010 0.9981619
#> 10: theta_bar    VA other 2010 1.0011126
#> 11: theta_bar    SC white 2010 1.0167582
#> 12: theta_bar    VA white 2010 1.0183112