print method for dgo_fit-class objects

print method for dgo_fit-class objects get_elapsed_time: extract chain run times from dgo_fit-class objects summary method for dgo_fit-class objects summarize method for dgo_fit-class objects as.data.frame method for dgo_fit-class objects rhats: extract split R-hats from dgo_fit-class objects

# S4 method for dgo_fit
show(object)

# S4 method for dgo_fit
print(x, ...)

print.dgo_fit(x, ...)

# S4 method for dgo_fit
get_elapsed_time(object, ...)

# S4 method for dgo_fit
summary(object, ..., verbose = FALSE)

# S4 method for dgo_fit
get_posterior_mean(object, pars = "theta_bar", ...)

summarize(x, ...)

# S4 method for dgo_fit
summarize(x, pars = "theta_bar", funs = c("mean", "sd",
  "median", "q_025", "q_975"))

# S3 method for dgo_fit
as.data.frame(x, ..., pars = "theta_bar",
  keep.rownames = FALSE)

rhats(x, ...)

# S4 method for dgo_fit
rhats(x, pars = "theta_bar")

Arguments

object
A dgo_fit-class object

x
A dgo_fit-class object

...
Further arguments to stanfit-class methods.

verbose
Whether to show the full output from the rstan method.

pars
Parameter name(s)

funs
Quoted names of summary functions. `q_025` is accepted as shorthand for `function(x) quantile(x, .025)`, and similarly `q_975`.

keep.rownames
Whether to retain original parameter names with numeric indexes, as output from RStan.

Value

A table giving split R-hats for model parameters

Examples

data(toy_dgirtfit)
summarize(toy_dgirtfit)

#>         param state race3 year       mean        sd      median      q_025
#>  1: theta_bar    SC black 2009  2.4845339 1.8374100  2.00633812  0.5105649
#>  2: theta_bar    SC black 2010  1.2314373 0.8527622  1.00416269  0.3030102
#>  3: theta_bar    SC other 2009 -1.4471515 1.2665559 -1.15791152 -4.9357601
#>  4: theta_bar    SC other 2010  0.0342708 0.7452419  0.05742328 -1.3407581
#>  5: theta_bar    SC white 2009 -1.7011866 1.1545385 -1.39596719 -4.7118286
#>  6: theta_bar    SC white 2010 -1.4465855 1.0472065 -1.18532541 -4.2231057
#>  7: theta_bar    VA black 2009  2.1251866 1.4580979  1.73325734  0.5150094
#>  8: theta_bar    VA black 2010  1.5124091 0.9533186  1.26516000  0.4000070
#>  9: theta_bar    VA other 2009 -1.0554578 1.0700110 -0.77739917 -4.1247751
#> 10: theta_bar    VA other 2010  0.3142603 0.5768678  0.26328872 -0.8119962
#> 11: theta_bar    VA white 2009 -0.9873349 0.7354829 -0.79534878 -2.8835486
#> 12: theta_bar    VA white 2010 -0.8153761 0.6536290 -0.65621335 -2.6393657
#>           q_975
#>  1:  7.23190297
#>  2:  3.63359289
#>  3:  0.02390292
#>  4:  1.46087840
#>  5: -0.32516265
#>  6: -0.24882620
#>  7:  6.52903753
#>  8:  3.85695569
#>  9:  0.24073128
#> 10:  1.68057822
#> 11: -0.14038527
#> 12: -0.08867862

data(toy_dgirtfit)
# access posterior samples
as.data.frame(toy_dgirtfit, pars = 'theta_bar')

#>           param state race3 year iteration      value
#>    1: theta_bar    SC black 2009         1  2.5349865
#>    2: theta_bar    SC black 2009         2  0.7095619
#>    3: theta_bar    SC black 2009         3  0.4745461
#>    4: theta_bar    SC black 2009         4  1.0656432
#>    5: theta_bar    SC black 2009         5  1.3278823
#>   ---                                                
#> 9596: theta_bar    VA white 2010       796 -1.3025842
#> 9597: theta_bar    VA white 2010       797 -0.7901599
#> 9598: theta_bar    VA white 2010       798 -1.0499463
#> 9599: theta_bar    VA white 2010       799 -0.5825237
#> 9600: theta_bar    VA white 2010       800 -0.5272909

data(toy_dgirtfit)
rhats(toy_dgirtfit)

#> Error in summary(x, par = pars, verbose = TRUE)$summary: $ operator is invalid for atomic vectors